refactor: 우승자 처리로직 위치 변경
