+++ src/main/java/racingcar/RacingCarDriver.java
@@ -8,7 +8,8 @@ import java.util.Random;
 
 public class RacingCarDriver {
 
-    private List<RacingCar> racingCars;
+    private final List<RacingCar> racingCars;
+    private List<String> winners = new ArrayList<>();
 
     public RacingCarDriver(List<RacingCar> racingCars) {
         if (racingCars != null) {
@@ -51,5 +52,46 @@ public class RacingCarDriver {
 
     public List<RacingCar> getRacingCars() {
         return racingCars;
+    }
+
+
+    /**
+     * 우승자가 누구인지 가져옵니다.
+     * @return
+     */
+    public List<String> getWinner() {
+        int longestDistance = 0;
+        for (RacingCar racingCar : this.getRacingCars()) {
+            int distance = racingCar.getDistance();
+            longestDistance = this.validateLongestDistance(longestDistance, distance);
+            this.addLongestDistance(longestDistance, distance, racingCar.getName());
+        }
+        return this.winners;
+    }
+
+    /**
+     * 주어진 거리가 가장 긴 거리이면 새로운 우승자 명단을 만들고 가장 긴 거리를 변경합니다.
+     * @param longestDistance
+     * @param distance
+     * @return longestDistance
+     */
+    private int validateLongestDistance(int longestDistance, int distance) {
+        if(distance > longestDistance) {
+            this.winners = new ArrayList<>();
+            longestDistance = distance;
+        }
+        return longestDistance;
+    }
+
+    /**
+     * 주어진 거리가 가장 긴 거리이면 우승자 명단에 이름을 추가합니다.
+     * @param longestDistance
+     * @param distance
+     * @param name
+     */
+    private void addLongestDistance(int longestDistance, int distance, String name) {
+        if(distance >= longestDistance) {
+            this.winners.add(name);
+        }
     }
 }
