+++ src/main/java/racingcar/ResultView.java
@@ -7,9 +7,6 @@ public class ResultView {
 
     private static final String RESULT_DASH = "-";
 
-    private List<String> winners = new ArrayList<>();
-
-
     /**
      * 주어진 거리만큼 '-'를 출력합니다.
      * @param distance
@@ -50,7 +47,7 @@ public class ResultView {
      * @param racingCarDriver
      */
     public void printWinner(RacingCarDriver racingCarDriver) {
-        List<String> winners = this.getWinner(racingCarDriver);
+        List<String> winners = racingCarDriver.getWinner();
         for (int i = 0; i < winners.size(); i++) {
             this.printDelimiter(i);
             System.out.print(winners.get(i));
@@ -68,45 +65,4 @@ public class ResultView {
         }
     }
 
-    /**
-     * 우승자가 누구인지 가져옵니다.
-     * @param racingCarDriver
-     * @return
-     */
-    private List<String> getWinner(RacingCarDriver racingCarDriver) {
-        int longestDistance = 0;
-        for (RacingCar racingCar : racingCarDriver.getRacingCars()) {
-            int distance = racingCar.getDistance();
-            longestDistance = this.validateLongestDistance(longestDistance, distance);
-            this.addLongestDistance(longestDistance, distance, racingCar.getName());
-
-        }
-        return this.winners;
-    }
-
-    /**
-     * 주어진 거리가 가장 긴 거리이면 새로운 우승자 명단을 만들고 가장 긴 거리를 변경합니다.
-     * @param longestDistance
-     * @param distance
-     * @return longestDistance
-     */
-    private int validateLongestDistance(int longestDistance, int distance) {
-        if(distance > longestDistance) {
-            this.winners = new ArrayList<>();
-            longestDistance = distance;
-        }
-        return longestDistance;
-    }
-
-    /**
-     * 주어진 거리가 가장 긴 거리이면 우승자 명단에 이름을 추가합니다.
-     * @param longestDistance
-     * @param distance
-     * @param name
-     */
-    private void addLongestDistance(int longestDistance, int distance, String name) {
-        if(distance >= longestDistance) {
-            this.winners.add(name);
-        }
-    }
 }
