+++ src/test/java/racingcar/RacingCarDriverTest.java
@@ -10,6 +10,7 @@ import racingcar.rule.RandomRacingRule;
 
 import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Collectors;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -61,5 +62,26 @@ public class RacingCarDriverTest {
 
         // then
         assertThat(distance).isEqualTo(Integer.parseInt(expected));
+    }
+
+    @ParameterizedTest
+    @CsvSource(value = {"3,2,2:test1", "3,2,3:test1,test3"}, delimiter = ':')
+    @DisplayName("우승자를 선별하는 테스트")
+    public void getWinner(String input, String expected) {
+        // given
+        List<Integer> distances = Arrays.stream(input.split(","))
+                                        .map(Integer::parseInt)
+                                        .collect(Collectors.toList());
+        List<RacingCar> sourceRacingCars
+                = Arrays.asList(new RacingCar("test1", distances.get(0))
+                                , new RacingCar("test2", distances.get(1))
+                                , new RacingCar("test3", distances.get(2)));
+        RacingCarDriver racingCarDriver = new RacingCarDriver(sourceRacingCars);
+
+        // when
+        String winner = racingCarDriver.getWinner().toString();
+
+        // then
+        assertThat(winner).contains(expected.split(","));
     }
 }
